tsmigrate: notification
