+++ src/add.tsx
@@ -2,7 +2,7 @@ import { Component, CSSProperties } from "react";
 
 import { Redirect } from "react-router-dom";
 import { addArticle } from "./api";
-import { addNotification } from "./store";
+import { addNotification } from "./notifications";
 import { apiRoot } from "./settings";
 import { waitFor } from "./utils";
