+++ src/notificationInterface.ts
@@ -1,10 +1,13 @@
 export type NotificationLevel = "default" | "error" | "info" | "success";
 
-export type Notification = {
+export type NotificationInit = {
 	data: JSX.Element | string;
-	time: number | null;
-	level: NotificationLevel;
-	closable: boolean;
-	context: any;
+	time?: number | null;
+	level?: NotificationLevel;
+	closable?: boolean;
+	context?: any;
+};
+
+export type Notification = NotificationInit & {
 	id?: number;
 };
