+++ src/notifications.tsx
@@ -1,4 +1,4 @@
-import { removeNotification } from "./store";
+import { setState, store } from "./store";
 import { Notification } from "./notificationInterface";
 
 type Props = {
@@ -6,7 +6,7 @@ type Props = {
 	className?: string;
 };
 
-export default function Notifications(props: Props) {
+export function Notifications(props: Props) {
 	return (
 		<div className="notifications">
 			{props.notifications &&
@@ -35,4 +35,84 @@ export default function Notifications(props: Props) {
 					))}
 		</div>
 	);
+}
+
+let notificationId: number = 0;
+
+type DeferredNotification = (remove: () => void) => Partial<Notification>;
+
+function createNotification(
+	notification: string | Partial<Notification>
+): Notification {
+	if (typeof notification === "string") {
+		notification = {
+			data: <span dangerouslySetInnerHTML={{ __html: notification }} />,
+			time: 3000,
+			level: "default",
+		};
+	} else if (typeof notification.data === "string") {
+		notification.data = (
+			<span dangerouslySetInnerHTML={{ __html: notification.data }} />
+		);
+	}
+	notification.id = notificationId++;
+	notification = Object.assign(
+		{
+			context: null,
+			level: "default",
+			time: 3000,
+			closable: true,
+		},
+		notification
+	);
+	//inject key into react component to avoid misrendering
+	(notification.data as JSX.Element).key = notification.id;
+	return notification as Notification;
+}
+
+export function addNotification(
+	notification: DeferredNotification | string | Partial<Notification>
+): Notification {
+	if (typeof notification === "function") {
+		// fuckery with indirection
+		const n = {};
+		const remover = () => {
+			store.dispatch({
+				type: "removeNotification",
+				notification: n,
+			});
+		};
+		Object.assign(n, createNotification(notification(remover)));
+		notification = n;
+	} else {
+		notification = createNotification(notification);
+	}
+	store.dispatch({
+		type: "addNotification",
+		notification,
+	});
+	if (notification.time !== null) {
+		setTimeout(() => {
+			store.dispatch({
+				type: "removeNotification",
+				notification: notification as Partial<Notification>,
+			});
+		}, notification.time);
+	}
+	return notification as Notification;
+}
+
+export function removeNotification(notification: Notification): void {
+	store.dispatch({
+		type: "removeNotification",
+		notification,
+	});
+}
+
+export function removeNotificationsWithContext(context: any): void {
+	setState({
+		notifications: store
+			.getState()
+			.notifications.filter((n: Notification) => n.context !== context),
+	});
 }
