+++ src/store.tsx
@@ -25,7 +25,7 @@ export interface StateAction {
 
 export interface NotificationAction {
 	type: "addNotification" | "removeNotification";
-	notification: any;
+	notification: Partial<Notification>;
 }
 
 const rootReducer = (
@@ -61,86 +61,6 @@ export function setState(state: Partial<State>): void {
 	store.dispatch({
 		type: "setState",
 		state,
-	});
-}
-
-let notificationId: number = 0;
-
-type DeferredNotification = (remove: () => void) => Partial<Notification>;
-
-function createNotification(
-	notification: string | Partial<Notification>
-): Notification {
-	if (typeof notification === "string") {
-		notification = {
-			data: <span dangerouslySetInnerHTML={{ __html: notification }} />,
-			time: 3000,
-			level: "default",
-		};
-	} else if (typeof notification.data === "string") {
-		notification.data = (
-			<span dangerouslySetInnerHTML={{ __html: notification.data }} />
-		);
-	}
-	notification.id = notificationId++;
-	notification = Object.assign(
-		{
-			context: null,
-			level: "default",
-			time: 3000,
-			closable: true,
-		},
-		notification
-	);
-	//inject key into react component to avoid misrendering
-	(notification.data as JSX.Element).key = notification.id;
-	return notification as Notification;
-}
-
-export function addNotification(
-	notification: DeferredNotification | string | Partial<Notification>
-): Notification {
-	if (typeof notification === "function") {
-		// fuckery with indirection
-		const n = {};
-		const remover = () => {
-			store.dispatch({
-				type: "removeNotification",
-				notification: n,
-			});
-		};
-		Object.assign(n, createNotification(notification(remover)));
-		notification = n;
-	} else {
-		notification = createNotification(notification);
-	}
-	store.dispatch({
-		type: "addNotification",
-		notification,
-	});
-	if (notification.time !== null) {
-		setTimeout(() => {
-			store.dispatch({
-				type: "removeNotification",
-				notification,
-			});
-		}, notification.time);
-	}
-	return notification as Notification;
-}
-
-export function removeNotification(notification: Notification): void {
-	store.dispatch({
-		type: "removeNotification",
-		notification,
-	});
-}
-
-export function removeNotificationsWithContext(context: any): void {
-	setState({
-		notifications: store
-			.getState()
-			.notifications.filter((n: Notification) => n.context !== context),
 	});
 }
